refactor: Rename package - `Member` -> `User`
