--- app/src/main/java/org/project/domain/member/domain/User.java
@@ -1,7 +1,0 @@
-package org.project.domain.member.domain;
-
-import org.springframework.stereotype.Repository;
-
-public class User {
-
-}
