+++ app/src/main/java/org/project/domain/user/domain/User.java
@@ -1,0 +1,5 @@
+package org.project.domain.user.domain;
+
+public class User {
+
+}
