+++ app/src/main/java/org/project/domain/user/repository/UserRepository.java
@@ -1,7 +1,6 @@
-package org.project.domain.member.repository;
+package org.project.domain.user.repository;
 
-import java.util.HashMap;
-import org.project.domain.member.domain.User;
+import org.project.domain.user.domain.User;
 import org.springframework.stereotype.Repository;
 
 @Repository
